Add geometryfinder and update findwinsows methods in site
