+++ IO.Astrodynamics.Net/IO.Astrodynamics/Body/CelestialItem.cs
@@ -65,6 +65,8 @@ public abstract class CelestialItem : ILocalizable, IEquatable<CelestialItem>
     /// </summary>
     protected DTO.CelestialBody ExtendedInformation;
 
+    private readonly GeometryFinder _geometryFinder = new GeometryFinder();
+
     /// <summary>
     /// Gets a value indicating whether the celestial item is a spacecraft.
     /// </summary>
@@ -363,7 +365,7 @@ public abstract class CelestialItem : ILocalizable, IEquatable<CelestialItem>
     {
         Func<Time, double> calculateDistance = date => { return GetEphemeris(date, observer, Frame.ICRF, aberration).ToStateVector().Position.Magnitude(); };
 
-        return FindWindowsWithCondition(searchWindow, calculateDistance, relationalOperator, value, stepSize);
+        return _geometryFinder.FindWindowsWithCondition(searchWindow, calculateDistance, relationalOperator, value, stepSize);
     }
 
     public IEnumerable<Window> FindWindowsOnOccultationConstraint(in Window searchWindow, ILocalizable observer,
@@ -380,7 +382,7 @@ public abstract class CelestialItem : ILocalizable, IEquatable<CelestialItem>
             return occultation == occultationType;
         };
 
-        return FindWindowsWithCondition(searchWindow, calculateOccultation, RelationnalOperator.Equal, true, stepSize);
+        return _geometryFinder.FindWindowsWithCondition(searchWindow, calculateOccultation, RelationnalOperator.Equal, true, stepSize);
     }
 
     public IEnumerable<Window> FindWindowsOnCoordinateConstraint(in Window searchWindow, ILocalizable observer, Frame frame,
@@ -448,150 +450,7 @@ public abstract class CelestialItem : ILocalizable, IEquatable<CelestialItem>
                 break;
         }
 
-        return FindWindowsWithCondition(searchWindow, evaluateCoordinates, relationalOperator, value, stepSize);
-    }
-
-    public bool EvaluateCondition<T>(T actualValue, T targetValue, RelationnalOperator operatorType)
-    {
-        if (actualValue is double actualDouble && targetValue is double targetDouble)
-        {
-            switch (operatorType)
-            {
-                case RelationnalOperator.Greater:
-                    return actualDouble > targetDouble;
-                case RelationnalOperator.GreaterOrEqual:
-                    return actualDouble >= targetDouble;
-                case RelationnalOperator.Lower:
-                    return actualDouble < targetDouble;
-                case RelationnalOperator.LowerOrEqual:
-                    return actualDouble <= targetDouble;
-                case RelationnalOperator.Equal:
-                    return System.Math.Abs(actualDouble - targetDouble) < 1e-9;
-                case RelationnalOperator.NotEqual:
-                    return System.Math.Abs(actualDouble - targetDouble) >= 1e-9;
-                default:
-                    throw new ArgumentException("Invalid operator");
-            }
-        }
-        else if (actualValue is bool actualBool && targetValue is bool targetBool)
-        {
-            switch (operatorType)
-            {
-                case RelationnalOperator.Equal:
-                    return actualBool == targetBool;
-                case RelationnalOperator.NotEqual:
-                    return actualBool != targetBool;
-                default:
-                    throw new ArgumentException("Invalid operator");
-            }
-        }
-        else
-        {
-            throw new ArgumentException("Invalid type");
-        }
-    }
-
-    public IEnumerable<Window> FindWindowsWithCondition<T>(Window searchWindow,
-        Func<Time, T> calculateValue,
-        RelationnalOperator relationalOperator, T targetValue,
-        TimeSpan stepSize)
-    {
-        List<Window> validWindows = new List<Window>();
-        var precision = TimeSpan.FromSeconds(1);
-        Time current = searchWindow.StartDate;
-        Time end = searchWindow.EndDate;
-
-        bool isInValidWindow = false;
-        Time? validWindowStart = null;
-
-        // Parcourir la plage de temps par stepSize
-        while (current <= end)
-        {
-            // Calculer la valeur actuelle à la date 'current'
-            T currentValue = calculateValue(current);
-
-            // Vérifier la condition à la date actuelle
-            if (EvaluateCondition<T>(currentValue, targetValue, relationalOperator))
-            {
-                // Si la condition est satisfaite et qu'on n'est pas déjà dans une fenêtre valide
-                if (!isInValidWindow)
-                {
-                    // Chercher précisément le début de la fenêtre avec une recherche binaire
-                    validWindowStart = FindTransition(current - stepSize, current, calculateValue, targetValue, relationalOperator, precision);
-                    isInValidWindow = true;
-                }
-            }
-            else
-            {
-                // Si la condition n'est plus satisfaite et qu'on était dans une fenêtre valide
-                if (isInValidWindow)
-                {
-                    // Chercher précisément la fin de la fenêtre avec une recherche binaire
-                    Time validWindowEnd = FindTransition(validWindowStart.Value, current, calculateValue, targetValue, ReverseOperator(relationalOperator), precision);
-
-                    // Ajouter la fenêtre précise à la liste
-                    validWindows.Add(new Window(validWindowStart.Value, validWindowEnd));
-
-                    // Réinitialiser le marqueur
-                    isInValidWindow = false;
-                    validWindowStart = null;
-                }
-            }
-
-            // Avancer à l'intervalle suivant
-            current = current.Add(stepSize);
-        }
-
-        // Gérer le cas où la condition est satisfaite jusqu'à la fin de la plage de recherche
-        if (isInValidWindow)
-        {
-            Time validWindowEnd = FindTransition(validWindowStart.Value, end, calculateValue, targetValue, ReverseOperator(relationalOperator), precision);
-            validWindows.Add(new Window(validWindowStart.Value, validWindowEnd));
-        }
-
-        return validWindows;
-    }
-
-    private Time FindTransition<T>(Time start, Time end, Func<Time, T> calculateValue,
-        T targetValue, RelationnalOperator relationalOperator, TimeSpan precision)
-    {
-        while ((end - start) > precision)
-        {
-            // Calculer le milieu de l'intervalle
-            Time mid = start + (end - start) / 2;
-
-            // Calculer la valeur au milieu
-            T midValue = calculateValue(mid);
-
-            // Vérifier la condition à la date 'mid'
-            if (EvaluateCondition<T>(midValue, targetValue, relationalOperator))
-            {
-                // Si la condition est satisfaite à 'mid', on continue à chercher vers le début pour trouver la transition exacte
-                end = mid;
-            }
-            else
-            {
-                // Sinon, on cherche vers la fin
-                start = mid;
-            }
-        }
-
-        // Retourner la date où la transition a été détectée avec la précision requise
-        return start;
-    }
-
-    private RelationnalOperator ReverseOperator(RelationnalOperator op)
-    {
-        return op switch
-        {
-            RelationnalOperator.Greater => RelationnalOperator.LowerOrEqual,
-            RelationnalOperator.Lower => RelationnalOperator.GreaterOrEqual,
-            RelationnalOperator.Equal => RelationnalOperator.NotEqual,
-            RelationnalOperator.LowerOrEqual => RelationnalOperator.Greater,
-            RelationnalOperator.GreaterOrEqual => RelationnalOperator.Lower,
-            RelationnalOperator.NotEqual => RelationnalOperator.Equal,
-            _ => throw new ArgumentOutOfRangeException()
-        };
+        return _geometryFinder.FindWindowsWithCondition(searchWindow, evaluateCoordinates, relationalOperator, value, stepSize);
     }
 
     #endregion
